templates/idle_clicker/requirements.txt
# Idle Clicker Bot Requirements
# Install with: pip install -r requirements.txt

# Core engine (install from parent project)
# cd ../.. && pip install -e .

# Required dependencies
python-dotenv>=1.0.0
aiogram>=3.0.0

# Optional: For enhanced features
schedule>=1.2.0  # For periodic tasks

